src/emu/ioport.cpp: track ioport_port's insertion order into ioport_list
And output it in frontend/mame/infoxml.cpp (`-listxml` CLI option). For
the rationale, see the added comment in src/emu/ioport.h.

Unlike the parent commit (which has been proposed to upstream and
withdrawn in [1]), an equivalent of the changes introduced by this
commit is not expected to ever be part of mainline MAME.

[1] https://github.com/mamedev/mame/pull/10486